Move unit tests to reflect new locations of internal classes
