+++ am-lib/src/test/java/uk/gov/hmcts/reform/amlib/internal/utils/PermissionsTest.java
@@ -1,11 +1,10 @@
-package uk.gov.hmcts.reform.amlib;
+package uk.gov.hmcts.reform.amlib.internal.utils;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import uk.gov.hmcts.reform.amlib.enums.Permission;
 import uk.gov.hmcts.reform.amlib.exceptions.UnsupportedPermissionsException;
-import uk.gov.hmcts.reform.amlib.internal.utils.Permissions;
 
 import java.util.Arrays;
 import java.util.HashSet;
@@ -21,7 +20,7 @@ import static uk.gov.hmcts.reform.amlib.enums.Permission.READ;
 import static uk.gov.hmcts.reform.amlib.enums.Permission.UPDATE;
 
 @SuppressWarnings("PMD")
-class PermissionTest {
+class PermissionsTest {
 
     @Test
     void sumOf_shouldCalculateSumOfPermissionsFromEmptyArray() {
